switch macos build to use gnu make
gnu make is required for order-only prerequisites to work, since they
rely on the pattern matching only matching files within a directory, not
within subdirectories of that directory - default apple make does not
seem to conform to this expectation, and so the build fails. using gmake
(gnu make) fixes this.